requirements.txt
pymupdf
pillow
pytesseract
openpyxl
python-docx
docx2pdf
streamlit
